Improve log_dhcp to always tell from which server a message came.
Currently this information is derived from the DHO_SERVERID option.
But when the server doesn't set it, no source info is logged. Print
the src ip of the udp packet as "from" in this case. This is not perfect
as there could be a dhcp forwarder involved, but this is better than nothing.

Thanks to Gerd von Egidy.