Add support for user defined options with embedded and encapsulated options. Install dhcpcd-embedded.conf to LIBEXECDIR which is read before dhcpcd.conf.
The idea is that this dhcpcd-embedded.conf is maintained only by me to support
RFC DHCP/DHCPv6 options that require known embedded or ncapsulated options
which are not easily added to our option structures in dhcp.c and dhcp6.c.
Eventually we may move the whole structure here so that a smaller binary
is produced and the definitions are easier to maintain.