dhcpcd: fix crash forcing lease renewals on inactive interfaces.
Summary:
Forcing a lease renewal on an interface assumes all interfaces
will have options. This is no longer the case because dhcpcd now
lists every interface on the system and the active flag indicates
whether dhcpcd is controlling it or not, so we need to check this
before blindly renewing.

While here, add a syslog entry to note the forced renew and make
the code a little more readable by returning early rather than
enclosing in if blocks.

Fixes T111.

Test Plan:
Start dhcpcd and send it SIGUSR1.
Check it does not crash and review renewal notice in syslog.

Reviewers: eiN5too0

Reviewed By: eiN5too0

Maniphest Tasks: T111

Differential Revision: https://dev.marples.name/D104