DHCPv6: deprecate addresses on stale leases.
Summary:
When a DHCPv6 Address or Prefix Delegation is leased, it may not be renewed as a
different one could be assigned. In this situation, we should
prompt the kernel to prefer this new one by deprecating the old
one. This is achieved by setting it's pltime to zero.

Related to T114.

Test Plan:
Configure your DHCPv6 server to assign a delegation from Pool A.
Configure dhcpcd to lease a Prefix delegation and assign it
to a downstream interface.
Start dhcpcd.
Configure your DHCPv6 server to asssign a delegation from Pool B.
Down/Up the link dhcpcd is using.
dhcpcd should assign the new delegation and depreate the old
by marking it's pltime as zero.

Reviewers: sthen, Harri

Reviewed By: Harri

Differential Revision: https://dev.marples.name/D108