DHCPv6: drop Prefix Delegations when dropping the lease.
Summary:
When dropping the lease, we should drop any Delegated Prefixes
as well, because they are from the lease.
Add a note to say if this needs to be changed, at least mark
the Delegated Prefixes as deprecated.

Fixes T114.

Test Plan:
Configure dhcpcd to request a Prefix Delegation and assign it to a downstream
interface.
Bring the interface down which dhcpcd requests the delegation from.
dhcpcd should remove the assigned delegation from the other interface.

Reviewers: Harri

Reviewed By: Harri

Maniphest Tasks: T114

Differential Revision: https://dev.marples.name/D111