dhcp6: implement lastlease and lastleaseextend
Summary:
This has resulted in a fair churn of code, but in summary:
  *  TIMEDOUT state introduced
  *  lease binding split out from handledata
  *  generic fail function introduced to attempt rebind
     of lastlease
  *  extended addresses are remembered and purged when a
     lease is re-negotiated
  *  defaults of 0 are no longer hard coded for most timers

Fixes T128.

Test Plan:
Obtain a dhcp6 lease, kill dhcpcd and the dhcp6 server.
run with `--lastlease` and check it's applied correctly.
Also, observe timers to ensure it renews, rebinds and expires.
Do same, but now with `--lastleaseextend as well`.
Observe same timers, but it should not expire but become an
infinite lease until a new one appears.

Reviewers: shahid

Reviewed By: shahid

Maniphest Tasks: T128

Differential Revision: https://dev.marples.name/D125