dhcp: arp announce existing address before reboot
This updates upstream ARP tables to send their ACK to the
interface we want to receive the message on and not one
which is not in use but has the same IP address.

Doing this means we no longer need to open DHCP sockets for
interfaces we're not interested in.