DHCP: Don't listen on the unspecified address when not in master mode
Otherwise we will swallow messages for another interface.
If we fail to bind to the address:BOOTPC socket then re-open
BPF. This could be due to address sharing (which isn't supported
withoutn master mode) or even worse another DHCP client running
on the same address. In this case, re-open BPF so we can listen
to RECONFIGURE messages.