privsep: Use another eloop instead of a blocking read.
We need to read the error status from making a privileged request.
This happens in an eloop already, but the read itself needs to block.
So that we can still be interupted, use an inner eloop whose sole
purpose is to monitor the root fd for message back. Then we can
make the fd non blocking.

dhcpcd should now exit cleanly once more.