privsep: Set a zero length receive buffer for write only sockets
We cannot use shutdown(2) because they are not connected.
Constantly draining would be a waste of CPU time, so just let
the buffer overflow. To ease the kernel as much as we can, set
a zero length buffer.

The kernel may still allocate a small buffer, but this is kernel
dependant and we're just trying to be helpful.