dhcpcd: Setup a socketpair in the launcher to write to stderr
Rather than duping stderr down to the processes.
This allows us to reopen stdout and stderr onto /dev/null right
away and means only the launcher process writes to stderr.

The downside is that any stdout from the script is now lost.
If that's needed, we could setup a stdout socketpair as well.