Fix handling of hostname argument
"hostname" argument (-h) is a simple string and should never be
interpreted as a hex string.  It is also kept as a null terminated
string, so when we parse something that is meant to be null terminated
string do not attempt to parse it as a HW addressa.  The same applies to
"script" (-c) argument.

Signed-off-by: Andrzej Ostruszka <amo@semihalf.com>