privsep: Don't find processes we just asked to stop
We could rebind a lease, get a NAK and enter DISCOVER.
We need to restart the BPF in the middle as the BPF filter could
change.

As such, add a started flag to each privsep process and when
searching for a process by id only find started ones.
If we ask them to stop then the started flag is removed.

Fixes errors about writing to stopping processes an unknown
processes exiting.