Replace uint16_t* with uint8_t* in in_cksum
Currently if we compile dhcpcd with -Os, we will get a checksum
error upon receiving packets from the DHCP server. This
prevents the client from binding a DHCP lease.

This checksum error comes from the checksums_valid() function volating
the strict aliasing. In this function, |pip| is a union representing
the pseudo IP header. It consists of |ip| of struct ip and |w| of
uint16_t[]. |ip| is assigned with the IP header data, but |w| is
actually used to calculate the checksum. As struct ip and uint16_t[] are
incompatible types, according to the strict aliasing, the compiler
thinks the assignment to |ip| is unused, so it simply ignores the
assignment. This leads to the incorrect checksum calculation.

According to the strict aliasing rule, a char* alias with any object
pointer, so we can choose to use uint8_t* (unsigned char*) to calculate
the checksum. We replace uint16_t* with uint8_t* in in_cksum, and also
remove the usage of |w| in checksums_valid().